// src/Solid/lib/useProfile/useProfile.d.ts
import { Quad_Object } from 'rdflib/lib/tf-types';

export declare interface IProfile {
    addresses: Quad_Object[];
    all: Quad_Object[];
    emails: Quad_Object[];
    fullName: Quad_Object[];
    images: Quad_Object[];
    inbox: Quad_Object[];
    name: Quad_Object[];
    phoneNumbers: Quad_Object[];
    prefs: Quad_Object[];
    privateTypes: Quad_Object[];
    publicTypes: Quad_Object[];
    roles: Quad_Object[];
    sameAs: Quad_Object[];
    seeAlso: Quad_Object[];
    storage: Quad_Object[];
};